Fix broken airflow target
Our prometheus config wasn't looking at the prometheus.io/port annotation and therefore falling back to look at defined ports on the service. Prometheus was attempting to scrape both ports defined on the statsd service, only one of which exposes metrics. The other port if for ingesetion, which was always timing out and marked DOWN in the prometheus targets list. This commit should fix that. Also falling back to global 5s scrape.